Add ml.po to .gitignore
The file was removed based on a DMCA Takedown Notice we received
from GitHub on behalf of Wipro Limited, claiming that they are
the copyright owner of the infringing code.

The file was removed from the whole git history and the tree was force-pushed.
It was also added to .gitignore to prevent adding it back by accident.

No major damage done, there were only 17 translated strings in the file.

In case someone creates a new Malayalam translation, ml.po can be removed
from .gitignore and replaced with the new file.